backend/requirements.txt
# Python Backend Dependencies
# Core Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.9  # PostgreSQL
# sqlite3 is included in Python standard library

# Authentication & Security
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6

# AI/ML Libraries
opencv-python>=4.8.0
numpy>=1.24.0
torch>=2.1.0  # PyTorch for SSD model
torchvision>=0.16.0  # For COCO SSD models
mediapipe>=0.10.0
dlib>=19.24.0
face-recognition>=1.3.0
pillow>=10.1.0  # For image processing

# Video/Audio Processing
ffmpeg-python>=0.2.0
moviepy>=1.0.3
pyaudio>=0.2.14

# Utilities
python-dotenv>=1.0.0
httpx>=0.25.0
aiofiles>=23.2.0
pillow>=10.1.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0

# Logging & Monitoring
loguru>=0.7.2

